+++ lib/dialogs.ts
@@ -20,9 +20,6 @@ export class Dialog {
 	#buttons: IDialogButton[]
 	#severity?: IDialogSeverity
 
-	/** @deprecated */
-	#html?: string
-
 	constructor(
 		name: string,
 		text: string,
@@ -33,16 +30,6 @@ export class Dialog {
 		this.#text = text
 		this.#buttons = buttons
 		this.#severity = severity
-		this.#html = undefined
-	}
-
-	/**
-	 * @deprecated DO NOT USE! It will be removed in the near future!
-	 * @param html HTML content
-	 */
-	setHTML(html: string) {
-		this.#html = html
-		return this
 	}
 
 	/**
@@ -58,7 +45,6 @@ export class Dialog {
 				name: this.#name,
 				text: this.#text,
 				severity: this.#severity,
-				html: this.#html,
 			},
 		)
 		if (!result) {